couple bug fixes
git-svn-id: https://svn.atnf.csiro.au/difx@6719 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/vdifio/trunk/src/vdifmux.c
@@ -119,7 +119,7 @@ int configurevdifmux(struct vdif_mux *vm, int inputFrameSize, int inputFramesPer
 
 	if(nThread > sizeof(vm->goodMask)*8)
 	{
-		fprintf(stderr, "Error: configurevdifmux: cannot run vdifmux on more than %d threads; %d requested.\n", sizeof(vm->goodMask)*8, nThread);
+		fprintf(stderr, "Error: configurevdifmux: cannot run vdifmux on more than %d threads; %d requested.\n", (int)sizeof(vm->goodMask)*8, nThread);
 
 		return -3;
 	}
